move mesh functions to OmegaHMesh class
