+++ pythonAPI/omegah2csg/__init__.py
@@ -11,11 +11,6 @@ from .openmcGeometry import (
     create_z_plane,
     create_z_cylinder,
     create_openmc_surface,
-    get_edge_coefficients,
-    get_num_of_boundary_edges,
-    get_boundary_edge_ids,
-    get_face_connectivity,
-    get_all_geometry_info,
     create_openmc_geometry,
     read_edge_coefficients_from_file,
     read_face_connectivity_from_file,
@@ -39,13 +34,6 @@ __all__ = [
     "create_z_plane",
     "create_z_cylinder",
     "create_openmc_surface",
-    # geometry queries
-    "get_edge_coefficients",
-    "get_num_of_boundary_edges",
-    "get_boundary_edge_ids",
-    "get_face_connectivity",
-    "get_all_geometry_info",
-    # openmc
     "create_openmc_geometry",
     "read_edge_coefficients_from_file",
     "read_face_connectivity_from_file",
